requirements.txt
streamlit
pandas
mysql-connector-python
sqlalchemy
bcrypt
seaborn
matplotlib
